deployment/requirements.txt
streamlit>=1.30.0
torch>=2.2.0
scikit-learn==1.6.1
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
joblib>=1.3.0
datasets>=2.16.0
pyarrow>=14.0.0
huggingface_hub>=0.20.0
python-dotenv
watchdog
